+++ ReConstruct.UI.Views/app-state.fs
@@ -1,4 +1,4 @@
 ﻿namespace ReConstruct.UI.View
 
 module AppState =
-    let mutable Level: Option<float32> = None
+    let mutable Level: Option<int> = None
